Finish most outstanding todos.
  * Move some formalisms to `formalisms.md`
  * Simply the type system by eliminating items entirely
  * Rename flags to variables and permit more types
  * Create a syntax for overrides
  * Add randomized variables
  * Specify interaction between overrides and configurations
  * Add a never type, for panics
  * Many minor fixes and clarifications
  * Expand on conditional tags
  * Allow conditionals to refer to one another.

The result is a langauge that, although not perfect, ought to be good
enough to start serious development and work with. It is now far
better-specified than previously, and includes a lot more features and
careful specification. It arguably has lost some functionality around
items, requiring that to be farmed out to library code, but this may not
matter much in practice and the only real way to find out is through
use.

The only thing I feel is remaining before implementation should proceed
in earnest is a formal proof of the override/configuration system, to
make sure it works or figure out how it ought to be simplified to make
it work. In particular, nested instnatiations lead to a layer of
complexity that requires quite careful thought.